+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Fixed
 
 - Fix crash when remote player enters after a pvp battle is done
+- Fix ig.ENTITY.OneTimeSwitch not switching to off animation on remote
 
 ## [0.7.5] 2026-03-06